requirements.txt
Django>=5.0.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8
django-environ>=0.11.0
whitenoise>=6.6.0
gunicorn>=21.2.0
requests>=2.31.0
djangorestframework>=3.14.0
django-allauth>=0.57.0
django-cors-headers>=4.3.1
google-auth-oauthlib>=1.1.0
google-auth>=2.25.0

